Move plot functions to Makie Extension
