+++ Project.toml
@@ -5,7 +5,6 @@ version = "0.1.0"
 
 [deps]
 ArchGDAL = "c9ce4bd3-c3d5-55b8-8973-c0e20141b8c3"
-Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
 DimensionalData = "0703355e-b756-11e9-17c0-8b28908087d0"
 DiskArrayEngine = "2d4b2e14-ccd6-4284-b8b0-2378ace7c126"
 DiskArrayTools = "fcd2136c-9f69-4db6-97e5-f31981721d63"
@@ -14,8 +13,6 @@ Extents = "411431e0-e8b7-467b-b5e0-f676ba4f2910"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 GeoInterface = "cf35fbd7-0cd7-5166-be24-54bfbe79505f"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
-MakieCore = "20f20a25-4f0e-4fdf-b5d1-57303727442b"
-Observables = "510215fc-4207-5dde-b226-833fc4488ee2"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 Proj = "c94c279d-25a6-4763-9509-64d165bea63e"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
@@ -25,9 +22,11 @@ Zarr = "0a941bbe-ad1d-11e8-39d9-ab76183a1d99"
 
 [weakdeps]
 ArchGDAL = "c9ce4bd3-c3d5-55b8-8973-c0e20141b8c3"
+Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 
 [extensions]
 PyramidSchemeArchGDALExt = "ArchGDAL"
+PyramidSchemeMakieExt = "Makie"
 
 [compat]
 Aqua = "0.8"
@@ -42,8 +41,6 @@ Extents = "0.1.3"
 FillArrays = "1.11"
 GeoInterface = "1.3.4"
 Makie = "0.21.2"
-MakieCore = "0.8.2"
-Observables = "0.5.5"
 OffsetArrays = "1.14"
 Proj = "1.7.1"
 Statistics = "1.10"
