+++ src/PyramidScheme.jl
@@ -11,22 +11,17 @@ using DiskArrayEngine: DiskArrayEngine, GMDWop, InputArray, LocalRunner, MovingW
 using DiskArrayEngine: create_outwindows, engine
 using DiskArrays: DiskArrays
 #using YAXArrays: savecube
-using Zarr
-using YAXArrayBase
-const YAB = YAXArrayBase
+using YAXArrayBase: YAXArrayBase as YAB
 using YAXArrays: Cube, YAXArray, to_dataset, savedataset, setchunks, open_dataset
 using Zarr: zcreate, writeattrs
 using DimensionalData: DimensionalData as DD
 using DimensionalData.Dimensions: XDim, YDim
-using Extents
+using Extents: Extent, extent
 using FillArrays: Zeros
 using Proj
-using Makie: Axis, Colorbar, DataAspect, Figure, FigureAxisPlot, Observable, Relative
-using Makie: on, heatmap!, image!
-import MakieCore: plot, plot!
 using OffsetArrays
 
-using Statistics
+using Statistics: mean
 
 export Pyramid, buildpyramids
 
@@ -411,24 +406,7 @@ function selectlevel(pyramid, ext;target_imsize=(1024, 512))
 end
 
 
-"""
-    plot(pyramids)
-Plot a Pyramid. 
-This will plot the coarsest resolution level at the beginning and will plot higher resolutions after zooming into the plot.
-This is expected to be used with interactive Makie backends.
-"""
-function plot(pyramid::Pyramid;colorbar=true, kwargs...)
-    #This should be converted into a proper recipe for Makie but this would depend on a pyramid type.
-    fig = Figure()
-    lon, lat = DD.dims(parent(pyramid))
-    ax = Axis(fig[1,1], limits=(extrema(lon), extrema(lat)), aspect=DataAspect())
-    hmap = plot!(ax, pyramid;kwargs...)
-    if colorbar
-        Colorbar(fig[1,2], hmap, height = Relative(3.5 / 4))
-    end
-    ax.autolimitaspect = 1
-    FigureAxisPlot(fig, ax, hmap)
-end
+
 
 xkey(keyext) = DD.dim2key(DD.dims(keyext, XDim))
 ykey(keyext) = DD.dim2key(DD.dims(keyext, YDim))
@@ -447,42 +425,6 @@ function switchkeys(dataext, keyext)
     yk = ykey(keyext)
     nt = NamedTuple{(xk, yk)}((dataext.X, dataext.Y))
     Extent(nt)
-end
-
-function plot!(ax, pyramid::Pyramid;interp=false, kwargs...)#; rastercrs=crs(parent(pyramid)),plotcrs=EPSG(3857), kwargs...)
-    tip = levels(pyramid)[end-2][:,:]
-    #@show typeof(tip)
-    data = Observable{DD.AbstractDimMatrix}(tip)
-    xval = only(values(Extents.extent(pyramid, XDim)))
-    yval = only(values(Extents.extent(pyramid, YDim)))
-    rasdataext = Extent(X=xval, Y=yval)
-    rasext = extent(pyramid)
-    xk = xkey(rasext)
-    yk = ykey(rasext)
-    on(ax.scene.viewport) do viewport
-        limext = Extents.extent(ax.finallimits[])
-
-        datalimit = switchkeys(limext, rasext)
-
-        data.val = selectlevel(pyramid, datalimit, target_imsize=viewport.widths)
-        notify(data)
-    end
-    on(ax.finallimits) do limits
-        limext = Extents.extent(limits)
-        # Compute limit in raster projection
-        #trans = Proj.Transformation(plotcrs, rastercrs, always_xy=true)
-        #datalimit = trans_bounds(trans, limext)
-        datalimit = switchkeys(limext, rasext)
-        if Extents.intersects(rasdataext, limext)
-            rasdata = selectlevel(pyramid, datalimit, target_imsize=ax.scene.viewport[].widths)
-            # Project selected data to plotcrs
-            #data.val = Rasters.resample(rasdata, crs=plotcrs, method=:bilinear )
-            data.val = rasdata
-        end
-        notify(data)
-    end
-    #@show typeof(data)
-    hmap = heatmap!(ax, data; interpolate=interp, kwargs...)
 end
 
 """
